# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Eplus v 0.0.1",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "summary": "Mostrar usuarios",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "summary": "Crear un nuevo usuario",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "roleId",
                                    "tuition",
                                    "name",
                                    "lastName",
                                    "motherLastName",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "roleId": {
                                        "description": "ID del rol del usuario",
                                        "type": "integer"
                                    },
                                    "tuition": {
                                        "description": "Matrícula del usuario",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "Nombre del usuario",
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "description": "Apellido paterno del usuario",
                                        "type": "string"
                                    },
                                    "motherLastName": {
                                        "description": "Apellido materno del usuario",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Correo electrónico del usuario",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario",
                                        "type": "string"
                                    },
                                    "universityCareer": {
                                        "description": "Carrera universitaria del usuario",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario dado de alta exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/users/{userId}": {
            "get": {
                "summary": "Mostrar un usuario por su ID",
                "operationId": "2fc03ba6d92e0712966a5ce4a7d9eaeb",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario encontrado."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "summary": "Actualizar un usuario existente",
                "operationId": "2af19474a22fcb71ca0c134728c7a07a",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "roleId",
                                    "tuition",
                                    "name",
                                    "lastName",
                                    "motherLastName",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "roleId": {
                                        "description": "ID del rol del usuario",
                                        "type": "integer"
                                    },
                                    "tuition": {
                                        "description": "Matrícula del usuario",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "Nombre del usuario",
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "description": "Apellido paterno del usuario",
                                        "type": "string"
                                    },
                                    "motherLastName": {
                                        "description": "Apellido materno del usuario",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Correo electrónico del usuario",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario",
                                        "type": "string"
                                    },
                                    "universityCareer": {
                                        "description": "Carrera universitaria del usuario",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario editado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "summary": "Eliminar un usuario",
                "operationId": "76f92b9a1870022f8b5c1c8a899c8a44",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID del usuario a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Se ha eliminado el usuario."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        }
    }
}